refactor: Junit 적용
